test: divide utils.test.ts
