+++ packages/client/test/unit/utils.test.ts
@@ -1,79 +1,12 @@
 import * as utils from '../../src/utils/utils'
-import { until } from '../../src/utils/promises'
-import { uuid } from '../../src/utils/uuid'
 import { inspect, format, DEFAULT_INSPECT_OPTS } from '../../src/utils/log'
 
 describe('utils', () => {
-    describe('uuid', () => {
-        it('generates different ids', () => {
-            expect(uuid('test')).not.toEqual(uuid('test'))
-        })
-        it('includes text', () => {
-            expect(uuid('test')).toContain('test')
-        })
-        it('increments', () => {
-            const uid = uuid('test') // generate new text to ensure count starts at 1
-            expect(uuid(uid) < uuid(uid)).toBeTruthy()
-        })
-    })
-
     describe('getEndpointUrl', () => {
         it('works', () => {
             const streamId = 'x/y'
             const url = utils.getEndpointUrl('http://example.com', 'abc', streamId, 'def')
             expect(url.toLowerCase()).toBe('http://example.com/abc/x%2fy/def')
-        })
-    })
-
-    describe('until', () => {
-        it('works with sync true', async () => {
-            const condition = jest.fn(() => true)
-            await until(condition)
-            expect(condition).toHaveBeenCalledTimes(1)
-        })
-
-        it('works with async true', async () => {
-            const condition = jest.fn(async () => true)
-            await until(condition)
-            expect(condition).toHaveBeenCalledTimes(1)
-        })
-
-        it('works with sync false -> true', async () => {
-            let calls = 0
-            const condition = jest.fn(() => {
-                calls += 1
-                return calls > 1
-            })
-            await until(condition)
-            expect(condition).toHaveBeenCalledTimes(2)
-        })
-
-        it('works with sync false -> true', async () => {
-            let calls = 0
-            const condition = jest.fn(async () => {
-                calls += 1
-                return calls > 1
-            })
-            await until(condition)
-            expect(condition).toHaveBeenCalledTimes(2)
-        })
-
-        it('can time out', async () => {
-            const condition = jest.fn(() => false)
-            await expect(async () => {
-                await until(condition, 100)
-            }).rejects.toThrow('Timeout')
-            expect(condition).toHaveBeenCalled()
-        })
-
-        it('can set interval', async () => {
-            const condition = jest.fn(() => false)
-            await expect(async () => {
-                await until(condition, 100, 20)
-            }).rejects.toThrow('Timeout')
-            expect(condition.mock.calls.length).toBeLessThan(7)
-            // ideally it should be 5.
-            expect(condition.mock.calls.length).toBeGreaterThan(4)
         })
     })
